+++ src/parser/roc_video_parser.cpp
@@ -26,6 +26,8 @@ RocVideoParser::RocVideoParser() {
     pic_count_ = 0;
     pic_width_ = 0;
     pic_height_ = 0;
+    bit_depth_luma_minus8_ = 0;
+    bit_depth_chroma_minus8_ = 0;
     new_seq_activated_ = false;
     frame_rate_.numerator = 0;
     frame_rate_.denominator = 0;
@@ -58,7 +60,7 @@ rocDecStatus RocVideoParser::Initialize(RocdecParserParams *pParams) {
         return ROCDEC_NOT_INITIALIZED;
     }
     // Initialize callback function pointers
-    pfn_sequece_cb_         = pParams->pfn_sequence_callback;     /**< Called before decoding frames and/or whenever there is a fmt change */
+    pfn_sequence_cb_         = pParams->pfn_sequence_callback;     /**< Called before decoding frames and/or whenever there is a fmt change */
     pfn_decode_picture_cb_  = pParams->pfn_decode_picture;        /**< Called when a picture is ready to be decoded (decode order)         */
     pfn_display_picture_cb_ = pParams->pfn_display_picture;       /**< Called whenever a picture is ready to be displayed (display order)  */
     pfn_get_sei_message_cb_ = pParams->pfn_get_sei_msg;           /**< Called when all SEI messages are parsed for particular frame        */
@@ -73,14 +75,6 @@ rocDecStatus RocVideoParser::Initialize(RocdecParserParams *pParams) {
     return ROCDEC_SUCCESS;
 }
 
-rocDecStatus RocVideoParser::MarkFrameForReuse(int pic_idx) {
-    if (pic_idx < 0) {
-        return ROCDEC_INVALID_PARAMETER;
-    }
-    //todo::
-    return ROCDEC_NOT_IMPLEMENTED;
-}
-
 void RocVideoParser::InitDecBufPool() {
     for (int i = 0; i < dec_buf_pool_size_; i++) {
         decode_buffer_pool_[i].use_status = kNotUsed;
@@ -92,6 +86,7 @@ void RocVideoParser::InitDecBufPool() {
 
 void RocVideoParser::CheckAndAdjustDecBufPoolSize(int dpb_size) {
     int min_dec_buf_pool_size = dpb_size + (parser_params_.max_display_delay > DECODE_BUF_POOL_EXTENSION ? parser_params_.max_display_delay : DECODE_BUF_POOL_EXTENSION);
+    // If DPB size decreases, we keep the existing pool and skip reconfiguration.
     if ( dec_buf_pool_size_ < min_dec_buf_pool_size) {
         dec_buf_pool_size_ = min_dec_buf_pool_size;
         decode_buffer_pool_.resize(dec_buf_pool_size_, {0});