+++ src/parser/vp9_parser.h
@@ -90,6 +90,12 @@ protected:
     DecodedPictureBuffer dpb_buffer_;
     Vp9Picture curr_pic_;
 
+    // Current surface size, used to support size change (down) on inter frames, where we keep the
+    // previously allocated surfaces and use them to store the smaller images.
+    uint32_t curr_surface_width_;
+    uint32_t curr_surface_height_;
+    uint32_t reconfig_option_;
+
     /*! \brief Function to parse one picture bit stream received from the demuxer.
      * \param [in] p_stream A pointer of <tt>uint8_t</tt> for the input stream to be parsed
      * \param [in] pic_data_size Size of the input stream
@@ -229,9 +235,9 @@ protected:
      * \param [in] p_stream Pointer to the bit stream
      * \param [inout] offset Bit offset
      * \param [out] p_uncomp_header Pointer to uncompressed header struct
-     *  \return None
+     * \return <tt>ParserResult</tt>
      */
-    void SegmentationParams(const uint8_t *p_stream, size_t &offset, Vp9UncompressedHeader *p_uncomp_header);
+    ParserResult SegmentationParams(const uint8_t *p_stream, size_t &offset, Vp9UncompressedHeader *p_uncomp_header);
 
     /*! \brief Function to parse probability syntax (read_delta_q(), 6.2.12)
      * \param [in] p_stream Pointer to the bit stream
@@ -244,9 +250,9 @@ protected:
      * \param [in] p_stream Pointer to the bit stream
      * \param [inout] offset Bit offset
      * \param [out] p_uncomp_header Pointer to uncompressed header struct
-     *  \return None
+     * \return <tt>ParserResult</tt>
      */
-    void TileInfo(const uint8_t *p_stream, size_t &offset, Vp9UncompressedHeader *p_uncomp_header);
+    ParserResult TileInfo(const uint8_t *p_stream, size_t &offset, Vp9UncompressedHeader *p_uncomp_header);
 
     /*! \brief Function to set up Y and UV dequantization values based on segmentation parameters. 8.6.1.
      *  \param [in] p_uncomp_header Pointer to uncompressed header struct