+++ CHANGELOG.md
@@ -2,11 +2,43 @@
 
 Full documentation for rocDecode is available at [https://rocm.docs.amd.com/projects/rocDecode/en/latest/](https://rocm.docs.amd.com/projects/rocDecode/en/latest/)
 
+## rocDecode 0.13.2 (unreleased)
+
+### Added
+
+* VP9 IVF container file parsing support in bitstream reader.
+* CTest for VP9 decode on bitstream reader.
+* HEVC/AVC/AV1/VP9 stream syntax error handling.
+* HEVC stream bit depth change handling and DPB buffer size change handling through decoder reconfiguration.
+* AVC stream DPB buffer size change handling through decoder reconfiguration.
+* rocDecode now uses the Cmake CMAKE_PREFIX_PATH directive.
+
+### Optimized
+
+* Decode session start latency reduction.
+* Bitstream type detection optimization in bitstream reader.
+
+### Resolved issues
+
+* Fixed a bug in picture files sample "videoDecodePicFiles" that can results in incorrect output frame count.
+* Fixed a decoded frame output issue in video size change cases.
+* Removed incorrect asserts of bitdepth_minus_8 in GetBitDepth() and num_chroma_planes in GetNumChromaPlanes() API calls in RocVideoDecoder utility class.
+
+### Removed
+
+* GetStream() interface call from RocVideoDecoder utility class
+
+### Changed
+
+* Changed asserts in query API calls in RocVideoDecoder utility class to error reports, to avoid hard stop during query in case error occurs and to let the caller decide actions.
+* `libdrm_amdgpu` is now explicitly linked with rocdecode.
+
 ## rocDecode 0.10.0 for ROCm 6.4
 
 ### Added
 
-* The new bitstream reader feature. The bitstream reader contains a few built-in stream file parsers, including elementary stream file parser and IVF container file parser. Currently the reader can parse AVC, HEVC and AV1 elementary stream files and AV1 IVF container files. More format support will be added in the future.
+* The new bitstream reader feature. The bitstream reader contains built-in stream file parsers, including an elementary stream file parser and an IVF container file parser. It can parse AVC/HEVC/AV1 elementary stream files and AV1 IVF container files. Additional format support can be added in the future.
+* VP9 decode support.
 * More CTests: VP9 test and tests on video decode raw sample.
 * Two new samples, videodecoderaw and videodecodepicfiles, have been added. videodecoderaw uses the bitstream reader instead of the FFMPEG demuxer to get picture data, and videodecodepicfiles shows how to decode an elementary video stream stored in multiple files with each file containing bitstream data of a coded picutre
 
@@ -20,6 +52,10 @@ Full documentation for rocDecode is available at [https://rocm.docs.amd.com/proj
 * FFMPEG executable requirement for the package
 
 ## rocDecode 0.8.0 for ROCm 6.3
+
+### Added
+
+* AV1 decode support
 
 ### Changed
 
@@ -76,6 +112,7 @@ Full documentation for rocDecode is available at [https://rocm.docs.amd.com/proj
 
 ### Additions
 
+* AVC decode support
 * FFMPEG V5.X Support
 * Mariner - Build Support
 
@@ -112,6 +149,7 @@ Full documentation for rocDecode is available at [https://rocm.docs.amd.com/proj
 
 ### Changes
 
+* Added HEVC decode support
 * Changed setup updates
 * Added AMDGPU package support
 * Optimized package dependencies